pyproject.toml
[project]
name = "lev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.2.1",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.120.1",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-core>=1.0.0",
    "langchain-groq>=1.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.50.0",
    "supabase>=2.24.0",
    "torch>=2.9.0",
    "uvicorn>=0.38.0",
]
